refactor: use module-based organization for framework-specific transport types
- Remove type prefixes (AxumSseServer, ActixSseServer, etc.)
- Organize implementations in framework-specific submodules (axum, actix_web)
- Each submodule exports types with consistent names (SseServer, StreamableHttpService)
- Update all imports to use new module paths
- Maintain backward compatibility with convenience aliases at module roots

This provides a cleaner API with better module organization and makes it easier
to add new framework implementations in the future.
